+++ app/include/view/media_filter.hpp
@@ -8,15 +8,27 @@
 
 class MediaFilter : public brls::Box {
 public:
-    MediaFilter(std::function<void(void)> cb);
+    MediaFilter();
     ~MediaFilter() override;
 
     bool isTranslucent() override { return true; }
+
+    brls::VoidEvent* getEvent() { return &this->event; }
 
     inline static int selectedSort = 1;
     inline static int selectedOrder = 1;
     inline static bool selectedPlayed = false;
     inline static bool selectedUnplayed = false;
+
+    inline static std::string sortList[] = {
+        "SortName",
+        "DateCreated",
+        "DatePlayed",
+        "PremiereDate",
+        "PlayCount",
+        "CommunityRating",
+        "Random",
+    };
 
 private:
     BRLS_BIND(brls::Box, cancel, "filter/cancel");
@@ -24,4 +36,6 @@ private:
     BRLS_BIND(brls::SelectorCell, sortOrder, "media/sort/order");
     BRLS_BIND(brls::BooleanCell, filterPlayed, "media/filter/played");
     BRLS_BIND(brls::BooleanCell, filterUnplayed, "media/filter/unplayed");
+
+    brls::VoidEvent event;
 };